+++ src/copy_this/modules/mo/mo_ogone/CHANGELOG.md
@@ -8,5 +8,6 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 - moved all files into module folder
+- flow theme is standard, azure tpl will be delivered if azure theme is active
 
 ## [3.3.1]